remove last video from context processor, call last_video template from base only if request.path equal /, call video tag in last_video template and add get_last_video in video_tags file
